+++ views/user/fields/add-citizenship.php
@@ -1,6 +1,7 @@
 <?php
 
 use app\widgets\buttons\Save;
+use app\widgets\buttons\Cancel;
 use kartik\select2\Select2;
 use yii\helpers\Html;
 use yii\widgets\ActiveForm;
@@ -26,14 +27,8 @@ ActiveForm::begin(); ?>
                 </div>
                 </div>
                 <div class="card-footer">
-                    <?= Html::submitButton(Yii::t('app', 'Save'), [
-                            'class' => 'btn btn-success',
-                            'title' => Yii::t('app', 'Save')]); ?>
-                    <?=  Save::widget(); ?>
-                    <?= Html::button(Yii::t('app', 'Cancel'), [
-                            'class' => 'btn btn-secondary',
-                            'data-dismiss' => 'modal',
-                            'title' => Yii::t('app', 'Cancel')]); ?>
+                    <?= Save::widget(); ?>
+                    <?= Cancel::widget(); ?>
                 </div>
             </div>
         </div>
