+++ widgets/buttons/Save.php
@@ -2,33 +2,22 @@
 
 namespace app\widgets\buttons;
 
-use app\widgets\base\SubmitButton;
+use yii\helpers\Html;
+use yii\base\Widget;
 use Yii;
 
-class Save extends SubmitButton
+class Save extends Widget
 {
 
     public function init()
     {
         parent::init();
-
-        $this->confirm = false;
-
-        $this->addClass .= 'btn btn-success';
-
-        $this->text = Yii::t('app', 'Save');
-
-        if ($this->addStyle == null) {
-            $this->addStyle['color'] = 'white';
-        } elseif (is_array($this->addStyle)) {
-            $this->addStyle['color'] = 'white';
-        } else {
-            $this->addStyle .= 'color: white;';
-        }
     }
 
     public function run()
     {
-        return parent::run();
+        return Html::submitButton(Yii::t('app', 'Save'), [
+            'class' => 'btn btn-success',
+            'title' => Yii::t('app', 'Save')]);
     }
 }
